+++ services/src/pro/layers/postgres_layer_db.rs
@@ -1,4 +1,4 @@
-use std::str::FromStr;
+use std::{collections::HashMap, str::FromStr};
 
 use async_trait::async_trait;
 use bb8_postgres::{
@@ -406,6 +406,7 @@ where
                         },
                         name: row.get(1),
                         description: row.get(2),
+                        properties: vec![],
                     })
                 } else {
                     CollectionItem::Collection(LayerCollectionListing {
@@ -415,6 +416,8 @@ where
                         },
                         name: row.get(1),
                         description: row.get(2),
+                        entry_label: None,
+                        properties: vec![],
                     })
                 }
             })
@@ -461,6 +464,8 @@ where
             description: row.get(1),
             workflow: serde_json::from_value(row.get(2)).context(error::SerdeJson)?,
             symbology: serde_json::from_value(row.get(3)).context(error::SerdeJson)?,
+            properties: vec![],
+            metadata: HashMap::new(),
         })
     }
 }